chore: optimize imports and move ExploreScreenTest.kt to appropriate package
- Optimized import
- Created explore package in androidTest under the corresponding ui package
- Moved ExploreScreenTest.kt to this new package
