+++ app/src/androidTest/java/com/android/unio/screen/EventListOverviewTest.kt
@@ -1,7 +1,10 @@
 package com.android.unio.ui.events
 
-import androidx.compose.ui.test.*
+import androidx.compose.ui.test.assertHasClickAction
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithTag
+import androidx.compose.ui.test.onNodeWithText
+import androidx.compose.ui.test.performClick
 import androidx.compose.ui.unit.ExperimentalUnitApi
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.unio.model.event.Event
